+++ src/_hooks/useFocusTrap.tsx
@@ -1,0 +1,39 @@
+import React from "react";
+
+export function useFocusTrap<T extends HTMLElement>(
+  element: React.RefObject<T>
+) {
+  React.useEffect(() => {
+    if (!element.current) return;
+
+    //get all focusable elements
+    const focusableElements = element.current.querySelectorAll(
+      'a[href], area[href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), button:not([disabled]), iframe, object, embed, [tabindex]:not([tabindex="-1"]), [contenteditable]'
+    );
+    const firstFocusableElement = focusableElements[0];
+    const lastFocusableElement =
+      focusableElements[focusableElements.length - 1];
+
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (e.key === "Tab") {
+        if (e.shiftKey) {
+          if (document.activeElement === firstFocusableElement) {
+            (lastFocusableElement as HTMLElement)?.focus();
+            e.preventDefault();
+          }
+        } else {
+          if (document.activeElement === lastFocusableElement) {
+            (firstFocusableElement as HTMLElement)?.focus();
+            e.preventDefault();
+          }
+        }
+      }
+    };
+
+    document.addEventListener("keydown", handleKeyDown);
+
+    return () => {
+      document.removeEventListener("keydown", handleKeyDown);
+    };
+  }, [element]);
+}